Trait pallet_confidential_docs::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type RemoveOrigin: EnsureOrigin<Self::Origin>;
    type MaxOwnedDocs: Get<u32>;
    type MaxSharedFromDocs: Get<u32>;
    type MaxSharedToDocs: Get<u32>;
    type DocNameMinLen: Get<u32>;
    type DocNameMaxLen: Get<u32>;
    type DocDescMinLen: Get<u32>;
    type DocDescMaxLen: Get<u32>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type RemoveOrigin: EnsureOrigin<Self::Origin>

type MaxOwnedDocs: Get<u32>

Maximum number of confidential documents that a user can own

type MaxSharedFromDocs: Get<u32>

Maximum number of confidential documents that a user can share

type MaxSharedToDocs: Get<u32>

Maximum number of confidential documents that can be shared to a user

type DocNameMinLen: Get<u32>

Minimum length for a document name

type DocNameMaxLen: Get<u32>

Maximum length for a document name

type DocDescMinLen: Get<u32>

Minimum length for a document description

type DocDescMaxLen: Get<u32>

Maximum length for a document description