Module pallet_confidential_docs::types

Defines the types required by the confidential docs pallet

Structs

Group

GroupMember

Group member

OwnedDoc

Owned confidential document

SharedDoc

Shared confidential document

Vault

User vault, the vault stores the cipher private key used to cipher the user documents. The way the user vault is ciphered depends on the login method used by the user

Enums

GroupRole

Group member role

Type Definitions

CID

Defines the type used by fields that store an IPFS CID

DocDesc

Defines the type used by fields that store a document description

DocName

Defines the type used by fields that store a document name

GroupName

Defines the type used by fields that store a group name

PublicKey

Defines the type used by fields that store a public key

UserId

Defines the type used by fields that store a UserId